Add main menu view
