+++ constants.py
@@ -1,0 +1,5 @@
+TILE_SPRITE_SCALING = 2
+PLAYER_SCALING = 0.2
+
+SCREEN_WIDTH = 960
+SCREEN_HEIGHT = 482
